refactor: Implement "Shallow but Focused" folder structure architecture
Major refactoring to simplify project structure while maintaining Single Responsibility Principle:

## 🏗️ Architectural Changes:
- Reduced folder depth from 4 levels to 2 levels
- Consolidated 33 files into 11 files (67% reduction)
- Maintained all existing functionality with improved organization

## 📁 New Structure:
- **src/analyzers/**: All analysis components (6 analyzers)
  - AudioExtractor, EmotionAnalyzer, SpeakerDiarizer, SpeechRecognizer
  - AcousticAnalyzer, VADProcessor
- **src/pipeline.py**: Complete pipeline architecture (647 lines)
- **src/models.py**: Unified data models (577 lines)
- **src/output.py**: All output processing (716 lines)
- **src/memory.py**: Memory optimization system (600+ lines)
- **src/utils.py**: Audio utilities and logging (700+ lines)

## 🔄 Key Consolidations:
1. **Pipeline Integration**: src/pipeline/ → pipeline.py
2. **Models Unification**: src/models/ → models.py
3. **Output Consolidation**: src/output/ + related files → output.py
4. **Memory Optimization**: utils/memory* → memory.py
5. **Utils Simplification**: utils/ → utils.py
6. **Analyzers Organization**: Individual files → src/analyzers/

## ✅ Benefits:
- **Improved Navigation**: Easier to find and understand code
- **Reduced Complexity**: Fewer nested directories and imports
- **Maintained SRP**: Each file <1000 lines with focused responsibility
- **Better Performance**: Simplified import paths and reduced overhead
- **Enhanced Maintainability**: Clear separation of concerns

## 🧪 Testing:
- All import paths updated and validated
- Pipeline build and execution tested successfully
- Existing functionality preserved and operational
